Use the existing django tox env to test rest_framework
