+++ .circleci/config.yml
@@ -296,15 +296,13 @@ jobs:
       - restore_cache:
           keys:
               - tox-cache-django-{{ checksum "tox.ini" }}
-      - run: tox -e '{py27,py34,py35,py36}-django{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached' --result-json /tmp/django.1.results
-      - run: tox -e '{py27,py34,py35,py36}-django-autopatch{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached' --result-json /tmp/django.2.results
-      - run: tox -e 'py36-django-restframework-djangorestframework' --result-json /tmp/django.3.results
+      - run: tox -e '{py27,py34,py35,py36}-django{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached-djangorestframework' --result-json /tmp/django.1.results
+      - run: tox -e '{py27,py34,py35,py36}-django-autopatch{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached-djangorestframework' --result-json /tmp/django.2.results
       - persist_to_workspace:
           root: /tmp
           paths:
             - django.1.results
             - django.2.results
-            - django.3.results
       - save_cache:
           key: tox-cache-django-{{ checksum "tox.ini" }}
           paths:
