+++ ddtrace/contrib/django/apps.py
@@ -1,4 +1,5 @@
 import logging
+import django
 
 # 3rd party
 from django.apps import AppConfig, apps
@@ -8,7 +9,6 @@ from .db import patch_db
 from .conf import settings
 from .cache import patch_cache
 from .templates import patch_template
-from .rest_framework import patch_rest_framework
 from .middleware import insert_exception_middleware, insert_trace_middleware
 
 from ...ext import AppTypes
@@ -70,9 +70,11 @@ class TracerConfig(AppConfig):
                 except Exception:
                     log.exception('error patching Django cache')
 
-            # Instrument rest_framework app to trace custom exception handling
-            if apps.is_installed('rest_framework'):
+            # Instrument rest_framework app to trace custom exception handling.
+            # rest_framework only supports django 1.10, 1.11 and 2.0.
+            if apps.is_installed('rest_framework') and django.VERSION >= (1, 10):
                 try:
+                    from .restframework import patch_rest_framework
                     patch_rest_framework(tracer)
                 except Exception:
                     log.exception('error patching rest_framework app')
