+++ tox.ini
@@ -27,9 +27,8 @@ envlist =
     {py27,py34,py35,py36}-elasticsearch{16,17,18,23,24,51,52,53,54}
     {py27,py34,py35,py36}-falcon{10,11,12}
     {py27,py34,py35,py36}-falcon-autopatch{10,11,12}
-    {py27,py34,py35,py36}-django{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
-    {py27,py34,py35,py36}-django-autopatch{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached
-    py36-django-restframework-djangorestframework
+    {py27,py34,py35,py36}-django{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached-djangorestframework
+    {py27,py34,py35,py36}-django-autopatch{18,19,110,111}-djangopylibmc06-djangoredis45-pylibmc-redis{210}-memcached-djangorestframework
     {py27,py34,py35,py36}-flask{010,011,012}-blinker
     {py27,py34,py35,py36}-flask-autopatch{010,011,012}-blinker
     {py27,py34,py35,py36}-flask{010,011,012}-flaskcache{013}-memcached-redis{210}-blinker
@@ -125,7 +124,6 @@ deps =
     django-autopatch111: django>=1.11,<1.12
     djangopylibmc06: django-pylibmc>=0.6,<0.7
     djangoredis45: django-redis>=4.5,<4.6
-    django-restframework: django>=1.11,<1.12
     djangorestframework: djangorestframework>=3.7
     flask010: flask>=0.10,<0.11
     flask011: flask>=0.11,<0.12
@@ -206,7 +204,6 @@ commands =
     elasticsearch{16,17,18,23,24,25,51,52,53,54}: nosetests {posargs} tests/contrib/elasticsearch
     django{18,19,110,111}: python tests/contrib/django/runtests.py {posargs}
     django-autopatch{18,19,110,111}: ddtrace-run python tests/contrib/django/runtests.py {posargs}
-    django-restframework: python tests/contrib/django/runtests_restframework.py {posargs}
     flaskcache{012,013}: nosetests {posargs} tests/contrib/flask_cache
     flask{010,011,012}: nosetests {posargs} tests/contrib/flask
     flask-autopatch{010,011,012}: ddtrace-run nosetests {posargs} tests/contrib/flask_autopatch
@@ -335,60 +332,49 @@ setenv =
     DATADOG_ENV = test
     DJANGO_SETTINGS_MODULE = app.settings_untraced
 
-[testenv:py27-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py27-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
 
-[testenv:py27-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py27-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py27-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py27-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py34-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py34-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py34-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py34-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py34-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py34-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py35-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py35-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py35-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py35-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py35-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py35-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py35-django-autopatch111-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py35-django-autopatch111-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py36-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py36-django-autopatch18-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py36-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py36-django-autopatch19-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py36-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py36-django-autopatch110-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-[testenv:py36-django-autopatch111-djangopylibmc06-djangoredis45-pylibmc-redis-memcached]
+[testenv:py36-django-autopatch111-djangopylibmc06-djangoredis45-pylibmc-redis-memcached-djangorestframework]
 setenv =
     {[django_autopatch]setenv}
-
-
-[django_restframework]
-setenv =
-    DATADOG_ENV = test
-    DJANGO_SETTINGS_MODULE = restframework.settings
-
-[testenv:py36-django-restframework-djangorestframework]
-setenv =
-    {[django_restframework]setenv}
-
 
 [flask_autopatch]
 setenv =
